Last chunk of data remaining
As part of the process, it can be there the scenario in which a
last chunk of data remains in the buffer. To cover that scenario
we add a last processing of counting words (if needed) just before
doing the calculation.
Also covered the posibility of non EOL caracter, so deal with this
scenario too.